fix(daemon): Re-think scheduling mechanism (Closes #28)
A new approach is a more simplified one. We take the neighbor node URL
and how many days we should live. Then we ping ourselves to
keep alive. Then the time comes to a limit. We ping the next node
so it starts and puts the logic on itself. If we see it's
okay then we turn off the daemon and hibernate