fix(): Fix wit.ai multi chunk response
Last update wit.ai has implemented a feature to transcribe the voice
message in form of chunks. now the result is slowly progressing and responds
with some partial transcriptions.
https://github.com/wit-ai/wit/issues/2206
To deal with this, we detect if axios was able to parse the result from the
string into object. If now, we are trying to fetch the last chunk and parse
data from there.