fix(WitAi): No final chunk means empty response
There is an edge case when the api response do not
have the chunk that marked as is_final. In this case
we used to say "there was an error during the voice
recognition". Technically it is not an error, so we
rather tell users it was an empty recognition. This
should not appear as an alert in the Loggly logs