fix(): Address Unsupported start tag error
Once we specified the markup to HTML, we brought more
opportunities to write messages. In the meantime we send
user name in the group chats. That led us to this error:
```
Bad Request: can't parse entities: Unsupported start tag \"|~foo_bar~|\" at byte offset 0"
```

Means telegram tries to parse the user name in the html
markup and fails. In this commit we disable HTML markup
for the recognised bot messages in groups. It always
contains raw text plus user name, so we are on a safe side