feat(threads): Support Forum groups
For Forum groups Telegram introduced more metadata in the Message object. Now we have
isTopicMessage flag that shows if we are in the thread group.
and also MessageThreadId is the anchor for the particular
thread in the group. To support this we need to
put the threadId into api payload every time we
want to answer in the thread, hence this change