feat(esm): Migrate to ESModules (Fix e2e tests)
The major pain is Jest tests. In fact, Jest works fine as
usual. The problem is hoisting. All imports must be located
on top of the module, even if we define it somewhere in the middle it will be bubbled up. In this sense jest.mock
does not work anymore. Jest team introduced a new API called
jest.unstable_mockModule to mock ES modules. But. Because
of hoisting we can not use it right away as well. The only
workaround it is to load app modules dynamically. Especially
in the e2e test where we run the whole application, we have
to load everything dynamically, hence we create these
dependency injectors, that can load a module in the test
after Jest mocked the data. this allows us to recover the
mocking behavior in tests without having to change the
test logic at all.