fix(err): Leave the chat if user restricted bot
When group admin restricts sending messages for the bot,
we constantly receive Telegram api errors with status 400
`Bad Request: not enough rights to send text messages to the chat`.
Since the bot has no other functionality, there is no reason
to stay in the group. THos flood our error log, so I
decided to leave the group if this happens.
In this commit we introduce the api reflector, which
detects the error type and acts like a middleware for
the error handlers. If we receive the message like I
mentioned above, the bor will leave the group