fix(telegram): Handle message is not modified
Since set language callback is asynchronous function, it
takes some time to resolve. During this time the user can
click on the button again. First attempt resolves and
second resolves as well. Hence we receive the error in
the logs. Although the user sees everything is fine. We
try to detect the message is not modified, and if so, make
it warning not error