fix(server): Link all promises
Initially it was an idea to release the connection as soon as possible. But this does nto show the correct metrics
in Newrelic dashboard. So now we are trying to connect all the promises in chain and see if it affects the service
runtime and error rate. if everything is okay, then this is the way to go.