feat(server): Migrate to Fastify
Fastify is significantly faster than ExpressJS. We are implementing a new
web router instance with Fastify and will run it under the feature flag.

In this commit we mostly implement the server instance. Since Sentry does
not provide good integration with Fastify, we add the sentry handlers as
well.
The health url updated to show the type fo the server instance. Also
fixes some issues with hanging handlers.