fix(lambda): Delete lambda implementation
I wanted to use Google Lambda function as a server-less solution for the bot initially.

FOr some reason the answer latency was off the charts so I got back to the classic server way.

If I to implement it again, most likely I would need to start from scratch. So no good reason to
keep it in the code base. I can always recover the git history