feat(languages): Prepare for new languages
In the current state the application is hard to equip with new languages.
It is doable but take a lot of effort to update here and there. I now
refactor the codebase in order to better support localisation and make
it seamless and flexible to add new languages. This required to separate
the locales into separate file, creating the loader and changing the way
how we get translation in the application.

I have also changed the way we load the VoiceConverter and how we supply
the configuration in such converters, so it will be easy to handle them
and add new converters in the future.