feat(whisper-alfa): Add Whisper processor
In this commit we establish the early version of the Whisper processor.
We need to check how to run it safely so it does not kill the process,
we also need to estimate the memory consumption is well below under the
limits of our low-resource availability model. Not to mention we want to
probably have a Whisper as a singleton and we need to somehow download
the recognition model - it take a lot of space so we can not just store
it in the repository

We do not enable Whisper as of now due to the challenges mentioned above,
But we also change the way we transform the voice file. Instead of
prism-media we use fluent-ffmpeg and we now always dump the wav file into
the file system as well. This is the requirement for the Whisper model
but I also want to unify the interface and stop having different
processing flows for audio and video. This change will be enabled with
the release