feat(languages): Explicit string interpolation in translations
We have to expose the interpolation api downstream to the consumers in
order to bring more customizations for the translation keys.
In the past we limited the list of variables in the translations due to
lack of usages and could embed the interpolation into translation module.
Now with subscriptions coming, we are expanding the approach so every
implementation can add the content in the translation string