feat(donate): Donate with Stars (#453)
Add the ability to use the Starts as a payment currency. This will be used
for donation but in the future we can expand the use cases if we find
extra.
I am adding the type for the PaymentChargeId so it is yet another identifier
in the world of Telegram.