feat(voice): Define two voice providers (#456)
In order to provide the premium feature, we want to use
two different voice providers. THe use case is - free users
use lets say WitAI and premium users get 11Labs or so, with
much better voice recognition.

We define the voice providers in the code now instead of
environment variables