fix(wit-ai): Validate the response payload (#459)
* fix(wit-ai): Validate the response payload

To make the app more reliable, we want to validate any
external data we receive. In this case we are validating
the result of the WitAI recognition

* fix(logs): Explicitly mark warnings to report

All warnings have flooded Sentry. And its okay to have such many
warning logs. It emphasises some major application component states.
But it is fine and we need to distinguish if it is something we really
want to look at

* fix(pnpm): Build dependencies