fix(deps): Bump dependencies, nuke ts-node
Since we only use eraseable types in typescript, we can easily get rig of
this ts-node thingy. The quickest idea is to use ts-blank-space loader
for the local development. This way we do not touch the production build.

We bump all dependencies as well and dynamically load the voice providers
just in case the node dependencies can take plenty of memory for no reason.