fix(WitAi): Fix memory leak (#473)
Axios error contains the whole api object with request, response and
the config. The config included the body payload, in our case it
can be a huge buffer object. apparently, in our implementation the NodeJS is having hard time to garbage collect it and hence
We hold Megabytes of unused buffers. In this commit we serialize
the axios error and then also delete the initial config, we already
know the place the axios is called so it is easy to navigate the error anyways