fix(): Break the import cycle (#477)
* fix(): Break the import cycle 1

in the telegram namespace we have a cycle types.ts <-> helpers.ts. We extract the model into separate file so solve this

* fix(): Break the import cycle 2

in the telegram namespace we have a cycle types.ts <-> text/index.ts. We extract the commands into separate file so solve this

* fix(): Break the import cycle 3

in the telegram namespace we have a cycle model.ts <-> helpers.ts. Put some helper closed to the actions. And move command helper to the command.ts module

* fix(deps): Bump dependencies

* fix(): Fix duplicated imports

* fix(): Fix duplicated imports